backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
httpx-socks==0.10.0
cryptography==44.0.0
apscheduler==3.10.4
supabase==2.11.0
loguru==0.7.3
tenacity==9.0.0
python-dotenv==1.0.1
PyJWT==2.10.1
